reformat `DetectionDataset` to `Florence2Dataset` and update deps
