+++ maestro/trainer/models/florence_2/metrics.py
@@ -5,7 +5,7 @@ import supervision as sv
 from PIL import Image
 from transformers import AutoProcessor
 
-from maestro.trainer.common.data_loaders.datasets import DetectionDataset
+from maestro.trainer.models.florence_2.loaders import Florence2Dataset
 
 DETECTION_CLASS_PATTERN = r"([a-zA-Z0-9 -]+)<loc_\d+>"
 
@@ -59,7 +59,7 @@ def process_output_for_text_metric(
     return predictions
 
 
-def get_unique_detection_classes(dataset: DetectionDataset) -> list[str]:
+def get_unique_detection_classes(dataset: Florence2Dataset) -> list[str]:
     class_set = set()
     for i in range(len(dataset)):
         _, suffix, _ = dataset[i]
